solving issue with normalizing server response
having a weird problem with convertFromRaw(). i was stubbing out my endpoints with a local flat file db, using convertToRaw() and using JSON.stringify() to send my POST request, and when calling that data with a GET request, using JSON.parse() and then convertFromRaw() and things were working WONDERFULLY. when i starting swapping out by stubbed out endpoints with real ones, i couldn’t get convertFromRaw() to work anymore. in manual testing i started to notice that the empty entityMap wasn’t being saved in my real DB. and that’s why convertFromRaw() stopped working. so my question is a. is this normal for a server to do this? i’m using Hapi and MongoDB and b. is there away to get convertFromRaw() to work without an entityMap? or should i just do a check of the server response to see if my object has that key?